+++ resources/js/Components/Reusable/NavItem.vue
@@ -7,7 +7,6 @@
         transition-all
         cursor-pointer
         hover:text-brand-red-500
-        text-gray-200
         "
   >
     <slot />
