+++ resources/js/Components/Slip/ShowSlipModal.vue
